refactor(signup): 약관 동의 관련 위젯 추출
